Add support for Rijndael algorithms (encryption methods 14, 15, and 16)
The code previously used PyCrypto's AES module to implement these
methods.  That did not work because they are actually Rijndael ciphers
and AES-256 is not the same as Rijndael-256 (and similarly with 128 and
192)¹.  It is possible to implement Rijndael-128 using PyCrypto's AES
module by setting the key_size to 32.  Unfortunately, it is not possible
to do so for Rijndael-192 or Rijndael-256.

To support all three Rijndael methods, the mcrypt module is needed.
Rather than use PyCrypto for just one of the Rijndael methods, replace
all three with an mcrypt version.

¹ http://stackoverflow.com/a/8355432